requirements.txt
requests
beautifulsoup4
pytest-playwright